requirements.txt
numpy
matplotlib
torch
torchvision
torchaudio
Pillow